+++ HISTORY.rst
@@ -13,6 +13,8 @@ Unreleased
 **Bugfixes**
 
 * ALCYN-1723: Fix delivery note csv (last column, tax and file encoding)
+* Improve the wizard to validate a picking and add the possibility to view all datas about a picking
+* Use sudo to remove stock.move when we need to delete an additional pack op
 
 **Build**